requirements.txt
absl-py==2.2.1
ale-py==0.10.2
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chex==0.1.89
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
dm-tree==0.1.9
docker-pycreds==0.4.0
etils==1.12.2
Farama-Notifications==0.0.4
filelock==3.18.0
flatbuffers==25.2.10
flax==0.10.4
fonttools==4.56.0
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.8.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
gymnasium==1.1.1
gymnasium-robotics==1.3.1
h5py==3.13.0
huggingface-hub==0.29.3
humanize==4.12.2
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
jax==0.5.3
jax-cuda12-pjrt==0.5.3
jax-cuda12-plugin==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
keras==3.9.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
minari==0.5.2
ml_collections==1.0.0
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
mujoco==3.1.6
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.5.5.64
opt_einsum==3.4.0
optax==0.2.4
optree==0.14.1
orbax-checkpoint==0.11.10
packaging==24.2
pandas==2.2.3
pettingzoo==1.24.3
pillow==11.1.0
platformdirs==4.3.7
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
pygame==2.6.1
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
sbx-rl==0.20.0
scipy==1.15.2
sentry-sdk==2.24.1
setproctitle==1.3.5
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
stable_baselines3==2.6.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.25.0
tensorstore==0.1.73
termcolor==2.5.0
tf_keras==2.19.0
toolz==1.0.0
torch==2.6.0
tqdm==4.67.1
treescope==0.1.9
triton==3.2.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
wandb==0.19.8
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
